+++ zga/zga.py
@@ -7,11 +7,11 @@ import shutil
 import subprocess
 import re
 import hashlib
-import json
 from Bio import SeqIO
 from zga import __version__
 from zga.assemblers import assemble
-
+from zga.essential import run_external, create_subdir
+from zga.preprocessing import read_processing
 
 def parse_args():
 	'''Returns argparse.Namespace'''
@@ -91,6 +91,8 @@ def parse_args():
 	reads_args.add_argument("--entropy-cutoff", type=float, default=-1,
 		help="Set between 0 and 1 to filter reads with entropy below "
 		+ "that value. Higher is more stringent. Default = -1, filtering disabled.")
+	reads_args.add_argument("--bbduk-k", type=int, default=19,
+		help="Kmer length used for finding contaminants with BBduk. [19]")
 	reads_args.add_argument("--bbduk-extra", nargs='*',
 		help="Extra options for BBduk. Should be space-separated.")
 	reads_args.add_argument("--tadpole-correct", action="store_true",
@@ -98,13 +100,8 @@ def parse_args():
 		+ "SPAdes correction may be disabled with \"--no-spades-correction\".")
 	reads_args.add_argument("--bbmerge", action="store_true",
 		help="Merge overlapped paired-end reads with BBMerge.")
-	reads_args.add_argument("--bbmerge-extend", type=int,
-		help="Perform k-mer read extension by specified length "
-		+ "if initial merging wasn't succesfull.")
-	reads_args.add_argument("--bbmerge-extend-kmer", type=int, default=40,
-		help="K-mer length for read extension, default 40.")
-	reads_args.add_argument("--bbmerge-trim", type=int,
-		help="Before merging trim bases with phred score less than a specified value.")
+	reads_args.add_argument("--bbmerge-extra", nargs='*',
+		help="Extra options for BBMerge. Should be space-separated.")
 	reads_args.add_argument("--normalize-kmer-cov", type=int,
 		help="Normalize read depth based on kmer counts to arbitrary value.")
 	reads_args.add_argument("--calculate-genome-size", action="store_true",
@@ -222,7 +219,7 @@ def parse_args():
 		else:
 			logger.error("File \"%s\" not found!", args.dfast_config)
 			raise FileNotFoundError(
-				"DFAST config file \"%s\" not found." % args.dfast_config
+				f"DFAST config file \"{args.dfast_config}\" not found."
 			)
 
 	return args
@@ -242,10 +239,10 @@ def check_reads(args):
 	logger.info("Checking input files.")
 
 	for reads in read_list:
-		for f in reads:
-			if f != "n/a" and not os.path.isfile(f):
-				logger.error("File %s doesn't exist", f)
-				raise FileNotFoundError("File %s doesn't exist" % f)
+		for f_name in reads:
+			if f_name != "n/a" and not os.path.isfile(f_name):
+				logger.error("File %s doesn't exist", f_name)
+				raise FileNotFoundError("File %s doesn't exist" % f_name)
 
 	short_libs = {}
 	if args.pe_1 and args.pe_2:
@@ -257,8 +254,8 @@ def check_reads(args):
 		short_libs["merged"] = args.pe_merged
 
 	if len(short_libs.values()) > 0:
-		short_lib_N = max(list(map(len, short_libs.values())))
-		for i in range(short_lib_N):
+		short_lib_count = max(list(map(len, short_libs.values())))
+		for i in range(short_lib_count):
 			libraries.append({"type": "short"})
 			for lib_type, lib in short_libs.items():
 				if len(lib) > i and lib[i] != "n/a":
@@ -293,39 +290,6 @@ def check_reads(args):
 	return libraries
 
 
-def create_subdir(parent, child) -> str:
-	'''Tries to create a subdirectory, returns path if succes.'''
-	path = os.path.join(parent, child)
-	try:
-		os.mkdir(path)
-	except Exception as e:
-		logger.critical("Impossible to create directory \"%s\"", path)
-		raise e
-	return path
-
-
-def run_external(args, cmd, keep_stdout=False, keep_stderr=True):
-	'''Run external command using subprocess.
-
-	Returns subprocess.CompletedProcess
-	'''
-	logger.debug("Running: %s", " ".join(cmd))
-	stderr_dest = subprocess.PIPE if keep_stderr else subprocess.DEVNULL
-	stdout_dest = subprocess.PIPE if keep_stdout else subprocess.DEVNULL
-
-	try:
-		r = subprocess.run(cmd, check=True, stderr=stderr_dest, stdout=stdout_dest, encoding="utf-8")
-		if args.transparent:
-			print(r.stderr, file=sys.stderr)
-		return r
-	except subprocess.CalledProcessError as e:
-		logger.error("Error during execution of: %s", ' '.join(e.cmd))
-		logger.info("Please see the logfile for additional information: %s",
-			os.path.join(args.output_dir, "zga.log"))
-		logger.debug("External tool stderr:\n%s", e.stderr)
-		return None
-
-
 def read_qc(args, reads):
 	'''Perform read QC with fastp'''
 	logger.info("Read quality control started")
@@ -333,299 +297,13 @@ def read_qc(args, reads):
 	precmd = ["fastp", "-L", "-Q", "-G", "-A", "-z", "1", "--stdout",
 		"-w", str(args.threads)]
 	for lib in reads:
-		for t, r in lib.items():
-			if t == "type":
+		for key, value in lib.items():
+			if key == "type":
 				continue
-			prefix = os.path.join(qcoutdir, os.path.split(r)[-1])
-			cmd = precmd + ["-i", r, "-h", f"{prefix}.html", "-j", f"{prefix}.json"]
-			logger.debug("QC of %s", r)
+			prefix = os.path.join(qcoutdir, os.path.split(value)[-1])
+			cmd = precmd + ["-i", value, "-h", f"{prefix}.html", "-j", f"{prefix}.json"]
+			logger.debug("QC of %s", value)
 			run_external(args, cmd)
-
-
-def remove_intermediate(path, *files):
-	'''Remove files from directory (path) and keep initial user files'''
-	for f in files:
-		if os.path.dirname(f) == path and os.path.exists(f):
-			logger.debug("Removing %s", f)
-			os.remove(f)
-
-
-def filter_by_tile(args, reads, readdir):
-	'''Run filterbytile.sh (BBmap) for Illumina read filtering'''
-	for index, lib in enumerate(reads, start=1):
-		if (lib["type"] == "short"
-			and "forward" in lib.keys()
-			and "reverse" in lib.keys()
-		):
-			initial = (lib['forward'], (lib['reverse']))
-			filtered_pe_r1 = os.path.join(readdir, f"lib{index}.filtered.r1.fq.gz")
-			filtered_pe_r2 = os.path.join(readdir, f"lib{index}.filtered.r2.fq.gz")
-
-			cmd = ["filterbytile.sh", f"in={initial[0]}", f"in2={initial[1]}",
-				f"out={filtered_pe_r1}", f"out2={filtered_pe_r2}",
-				f"-Xmx={args.memory_limit}G"]
-
-			if run_external(args, cmd) is not None:
-				remove_intermediate(readdir, *initial)
-				lib['forward'], lib['reverse'] = filtered_pe_r1, filtered_pe_r2
-			else:
-				logger.warning("Filtering by tile wasn't perfomed correctly.")
-
-	return reads
-
-
-def merge_bb(args, reads, readdir):
-	'''Performs merging (and extension) of overlapping paired-end reads.
-
-	Parameters:
-	args (argparse.Namespace)
-	reads (dict) : input reads
-	readdir (path) : path to reads output directory
-
-	Returns:
-	reads (dict) : modified if bbmerge returns 0.
-	'''
-	for index, lib in enumerate(reads, start=1):
-		if (lib["type"] == "short"
-			and "forward" in lib.keys()
-			and "reverse" in lib.keys()
-			and "merged" not in lib.keys()
-		):
-			# Initial, unmerged and merged filenames
-			initial = (lib['forward'], (lib['reverse']))
-			u1 = os.path.join(readdir, f"lib{index}.u1.fq")
-			u2 = os.path.join(readdir, f"lib{index}.u2.fq")
-			merged = os.path.join(readdir, f"lib{index}.merged.fq")
-
-			cmd = ["bbmerge.sh", f"Xmx={args.memory_limit}G", f"t={args.threads}",
-			f"in={initial[0]}", f"in2={initial[1]}",
-			f"outu1={u1}", f"outu2={u2}", f"out={merged}"]
-
-			if args.bbmerge_extend or bbmerge_extend_kmer:
-				cmd += [f"extend2={args.bbmerge_extend}",
-				f"k={args.bbmerge_extend_kmer}", "rsem=t"]
-			else:
-				cmd.append("strict=t")
-
-			if args.bbmerge_trim:
-				cmd += ["qtrim2=t", f"trimq={args.bbmerge_trim}"]
-			logger.info("Merging paired-end reads.")
-
-			if run_external(args, cmd) is not None:
-				remove_intermediate(readdir, *initial)
-				lib['forward'], lib['reverse'] = u1, u2
-				lib['merged'] = merged
-
-	return reads
-
-
-def repair_pair(args, readdir, lib, index):
-	'''Run repair.sh from BBmap
-
-	Returns:
-	(Fixed R1, Fixed R2), Singletons
-	'''
-	singletons = os.path.join(readdir, f"lib{index}.singletons.fq")
-	fixed = (os.path.join(readdir, f"lib{index}.repaired.r1.fq"),
-		os.path.join(readdir, f"lib{index}.repaired.r2.fq"))
-	cmd = ["repair.sh", f"in={lib[0]}", f"in2={lib[1]}",
-		f"out={fixed[0]}", f"out2={fixed[1]}", f"outs={singletons}",
-		f"Xmx={args.memory_limit}G"]
-	if run_external(args, cmd) is not None:
-		return (fixed, singletons)
-	logger.error("Error during repair of paired-end reads %s and %s",
-		lib[0], lib[1])
-	sys.exit(1)
-
-
-def bbduk_process(args, reads, readdir):
-	"""Perform trimming and filtering of short reads"""
-	bbduk_kmer = 19  # K-mer length for contaminant/adapter removal
-	precmd = ["bbduk.sh", f"Xmx={args.memory_limit}G", f"t={args.threads}",
-			f"ref={args.adapters}", f"k={bbduk_kmer}", "ktrim=r",
-			"qtrim=r", f"trimq={args.quality_cutoff}",
-			f"entropy={args.entropy_cutoff}",
-			f"minlength={args.min_short_read_length}"]
-	if args.bbduk_extra:
-		precmd += bbduk_extra
-
-	for index, lib in enumerate(
-		[lib for lib in reads if lib["type"] == "short"],
-		start=1):
-
-		if "forward" in lib.keys() and "reverse" in lib.keys():
-			logger.info("Trimming and filtering paired end reads")
-			initial = (lib['forward'], (lib['reverse']))
-			out_pe1 = os.path.join(readdir, f"lib{index}.r1.fq")
-			out_pe2 = os.path.join(readdir, f"lib{index}.r2.fq")
-			out_stats = os.path.join(readdir, f"lib{index}.bbduk.pe.txt")
-			cmd = precmd + [f"in={initial[0]}", f"in2={initial[1]}",
-				f"out={out_pe1}", f"out2={out_pe2}", f"stats={out_stats}"]
-
-			if run_external(args, cmd) is not None:
-				remove_intermediate(readdir, *initial)
-				lib['forward'], lib['reverse'] = out_pe1, out_pe2
-			else:
-				logger.error(
-					"Error during processing paired-end reads %s and %s",
-					initial[0], initial[1]
-				)
-				logger.warning("Trying to repair paired-end reads.")
-				fixed, discarded = repair_pair(args, readdir, initial, index)
-				remove_intermediate(readdir, *initial)
-				cmd = precmd + [f"in={fixed[0]}", f"in2={fixed[1]}",
-					f"out={out_pe1}", f"out2={out_pe2}", f"stats={out_stats}"]
-				if run_external(args, cmd) is not None:
-					remove_intermediate(readdir, *fixed)
-					lib['forward'], lib['reverse'] = out_pe1, out_pe2
-					if 'single' not in lib.keys():
-						lib['single'] = discarded
-
-		for read_type in ["single", "merged"]:
-			if read_type in lib.keys():
-				logger.info("Trimming and filtering %s reads", read_type)
-				initial = lib[read_type]
-				out = os.path.join(readdir, f"lib{index}.{read_type}.fq")
-				out_stats = os.path.join(readdir, f"lib{index}.bbduk.{read_type}.txt")
-				cmd = precmd + [f"in={initial}", f"out={out}",
-					f"stats={out_stats}"]
-
-				if run_external(args, cmd) is not None:
-					remove_intermediate(readdir, initial)
-					lib[read_type] = out
-
-	return reads
-
-
-def tadpole_correct(args, reads, readdir):
-	'''Correct short reads with tadpole'''
-	precmd = ["tadpole.sh",
-		f"Xmx={args.memory_limit}G",
-		f"t={args.threads}",
-		"mode=correct"]
-
-	for index, lib in enumerate([lib for lib in reads if lib["type"] == "short"], start=1):
-		if "forward" in lib.keys() and "reverse" in lib.keys():
-			logger.info("Error correction of paired end reads")
-			initial = (lib['forward'], (lib['reverse']))
-			out_pe1 = os.path.join(readdir, f"lib{index}.ecc.pe.r1.fq")
-			out_pe2 = os.path.join(readdir, f"lib{index}.ecc.pe.r2.fq")
-			cmd = precmd + [f"in={initial[0]}", f"in2={initial[1]}",
-				f"out={out_pe1}", f"out2={out_pe2}"]
-
-			if run_external(args, cmd) is not None:
-				remove_intermediate(readdir, *initial)
-				lib['forward'], lib['reverse'] = out_pe1, out_pe2
-
-		for read_type in ["single", "merged"]:
-			if read_type in lib.keys():
-				logger.info("Error correction of %s reads", read_type)
-				initial = lib[read_type]
-				out = os.path.join(readdir, f"lib{index}.ecc.{read_type}.fq")
-				cmd = precmd + [f"in={initial}", f"out={out}"]
-
-				if run_external(args, cmd) is not None:
-					remove_intermediate(readdir, initial)
-					lib[read_type] = out
-
-	return reads
-
-
-def bbnorm(args, reads, readdir):
-	'''Normalize reads by k-mer coverage depth with BBnorm'''
-	precmd = ["bbnorm.sh",
-		f"Xmx={args.memory_limit}G",
-		f"t={args.threads}",
-		f"minq={args.quality_cutoff}",
-		f"target={args.normalize_kmer_cov}"]
-
-	for index, lib in enumerate([lib for lib in reads if lib["type"] == "short"], start=1):
-		if "forward" in lib.keys() and "reverse" in lib.keys():
-			logger.info("Normalization of paired end reads")
-			initial = (lib['forward'], (lib['reverse']))
-			out_pe1 = os.path.join(readdir, f"lib{index}.norm.pe.r1.fq")
-			out_pe2 = os.path.join(readdir, f"lib{index}.norm.pe.r2.fq")
-			cmd = precmd + [f"in={initial[0]}", f"in2={initial[1]}",
-				f"out={out_pe1}", f"out2={out_pe2}"]
-
-			if run_external(args, cmd) is not None:
-				remove_intermediate(readdir, *initial)
-				lib['forward'], lib['reverse'] = out_pe1, out_pe2
-
-		for read_type in ["single", "merged"]:
-			if read_type in lib.keys():
-				logger.info("Normalization of %s reads", read_type)
-				initial = lib[read_type]
-				out = os.path.join(readdir, f"lib{index}.norm.{read_type}.fq")
-				cmd = precmd + [f"in={initial}", f"out={out}"]
-
-				if run_external(args, cmd) is not None:
-					remove_intermediate(readdir, initial)
-					lib[read_type] = out
-
-	return reads
-
-
-def compress_reads(args, reads, readdir):
-	'''Compress reads with pigz or gzip after processing'''
-
-	if shutil.which('pigz') is not None:
-		cmd = ['pigz', '-p', str(args.threads)]
-	else:
-		cmd = ['gzip']
-
-	for lib in reads:
-		for k, v in lib.items():
-			f = os.path.join(readdir, v)
-			if os.path.isfile(f) and os.path.splitext(f)[-1] not in ('gz', 'bz2'):
-				logger.debug("Compressing %s", v)
-				if run_external(args, cmd + [f]) is not None:
-					lib[k] = f'{v}.gz'
-
-	return reads
-
-
-def read_processing(args, reads):
-	'''Pipeline for read processing
-
-	Returns
-	reads (dict)
-	'''
-	logger.info("Reads processing started")
-	readdir = create_subdir(args.output_dir, "reads")
-	sr_adapters = os.path.join(
-		os.path.dirname(os.path.abspath(__file__)),
-		"data/sr.adapters.fasta")
-
-	if args.adapters and os.path.isfile(args.adapters):
-		args.adapters = os.path.abspath(args.adapters)
-	else:
-		args.adapters = sr_adapters
-
-	if args.filter_by_tile:
-		reads = filter_by_tile(args, reads, readdir)
-
-	reads = bbduk_process(args, reads, readdir)
-
-	if args.tadpole_correct:
-		reads = tadpole_correct(args, reads, readdir)
-
-	if args.normalize_kmer_cov:
-		reads = bbnorm(args, reads, readdir)
-
-	# Merging overlapping paired-end reads
-	if args.bbmerge or args.bbmerge_trim or bbmerge_extend:
-		reads = merge_bb(args, reads, readdir)
-
-	# Processing Illumina mate-pairs
-	if not args.no_nxtrim:
-		reads = mp_read_processing(args, reads, readdir)
-
-	reads = compress_reads(args, reads, readdir)
-
-	logger.info("Read processing finished")
-
-	return reads
 
 
 def mash_estimate(args, reads):
@@ -664,7 +342,7 @@ def mash_estimate(args, reads):
 		", ".join(reads_to_sketch))
 	r = run_external(args, cmd, keep_stderr=True)
 
-	if r is not None:
+	if r:
 		result = r.stderr.split("\n")
 		estimations = [float(x.split()[-1]) for x in result if "Estimated" in x.split()]
 		best_estimation = sorted(zip(estimations[::2], estimations[1::2]), key=lambda x: -x[1])[0]
@@ -674,38 +352,6 @@ def mash_estimate(args, reads):
 	else:
 		logger.error("Genome size estimation with \"mash\" failed.")
 		return None
-
-
-def mp_read_processing(args, reads, readdir):
-	'''Filtering Illumina mate pair reads'''
-	for index, lib in enumerate([lib for lib in reads if lib["type"] == "mate-pair"], start=1):
-		prefix = os.path.join(readdir, f"mate.{index}")
-
-		cmd = [
-			"nxtrim", "-1", lib['forward'], "-2", lib['reverse'],
-			"--separate", "--rf", "--justmp", "-O", prefix, "-l",
-			str(args.min_short_read_length)
-		]
-		logger.info("Processing mate-pair reads.")
-		if run_external(args, cmd) is not None:
-			if args.use_unknown_mp:
-				with open(f"{prefix}_R1.all.fastq.gz", "wb") as dest:
-					with open(f"{prefix}_R1.mp.fastq.gz", "rb") as src:
-						shutil.copyfileobj(src, dest)
-					with open(f"{prefix}_R1.unknown.fastq.gz", "rb") as src:
-						shutil.copyfileobj(src, dest)
-				with open(f"{prefix}_R2.all.fastq.gz", "wb") as dest:
-					with open(f"{prefix}_R2.mp.fastq.gz", "rb") as src:
-						shutil.copyfileobj(src, dest)
-					with open(f"{prefix}_R2.unknown.fastq.gz", "rb") as src:
-						shutil.copyfileobj(src, dest)
-				lib['forward'], lib['reverse'] = (
-					f"{prefix}_R1.all.fastq.gz", f"{prefix}_R2.all.fastq.gz")
-			else:
-				lib['forward'], lib['reverse'] = (
-					f"{prefix}_R1.mp.fastq.gz", f"{prefix}_R2.mp.fastq.gz")
-
-	return reads
 
 
 def map_short_reads(args, assembly, reads, target):
@@ -725,7 +371,7 @@ def map_short_reads(args, assembly, reads, target):
 		"-a", "-o", target, assembly, reads]
 
 	logger.info("Mapping reads: %s", reads)
-	if run_external(args, cmd) is not None:
+	if run_external(args, cmd):
 		return target
 	else:
 		logger.error("Unsuccesful mapping.")
@@ -751,17 +397,16 @@ def racon_polish(args, assembly, reads) -> str:
 				r = run_external(args, cmd, keep_stdout=True)
 				if os.path.exists(mapping):
 					os.remove(mapping)
-				if r is not None:
+				if r:
 					digest = hashlib.md5(r.stdout.encode('utf-8')).hexdigest()
 					suffix = "".join(
 						[chr(65 + (int(digest[x], 16) + int(digest[x + 1], 16)) % 26) for x in range(0, 20, 2)]
 					)
 					fname = os.path.join(polish_dir, f"polished.{suffix}.fna")
 					try:
-						handle = open(fname, 'w')
-						handle.write(r.stdout)
-						handle.close()
-						assembly = fname
+						with open(fname, 'w') as handle:
+							handle.write(r.stdout)
+							assembly = fname
 					except Exception:
 						logger.error("Error during polishing: impossible to write file %s", fname)
 			else:
@@ -787,7 +432,8 @@ def annotate(args) -> str:
 	logger.info("Genome annotation started")
 
 	try:
-		version_stdout = subprocess.run(["dfast", "--version"], encoding="utf-8",
+		version_stdout = subprocess.run(
+			["dfast", "--version"], encoding="utf-8", check=True,
 			stderr=subprocess.PIPE, stdout=subprocess.PIPE).stdout
 		version = re.search(r'ver. (\d\S*)', version_stdout)[1]
 		logger.debug("DFAST version %s available.", version)
@@ -839,7 +485,8 @@ def check_phix(args):
 	logger.debug("Running: %s", " ".join(cmd))
 	try:
 		blast_out = str(subprocess.run(
-			cmd, universal_newlines=True, stderr=subprocess.PIPE, stdout=subprocess.PIPE
+			cmd, universal_newlines=True, check=True,
+			stderr=subprocess.PIPE, stdout=subprocess.PIPE
 		).stdout).rstrip()
 	except Exception as e:
 		logger.error("Error during BLAST+ run")
@@ -896,7 +543,8 @@ def run_checkm(args):
 				encoding="utf-8",
 				universal_newlines=True,
 				stderr=subprocess.DEVNULL,
-				stdout=subprocess.PIPE
+				stdout=subprocess.PIPE,
+				check=True
 			).stdout
 		except Exception as e:
 			logger.critical("Failed to run CheckM!")
@@ -938,54 +586,11 @@ def run_checkm(args):
 	shutil.rmtree(checkm_indir)
 	shutil.rmtree(checkm_outdir)
 
-	if r is not None:
+	if r:
 		return checkm_outfile
 	else:
 		logger.error("CheckM didn't finish properly.")
 		return None
-
-
-def get_N_L_metric(lengths, value=50):
-	'''Returns a tuple containing NX and LX metric'''
-	l_total = sum(lengths)
-	metric = 0.01 * value
-	l_sum = 0
-	for i, x in enumerate(lengths, 1):
-		l_sum += x
-		if l_sum >= l_total * metric:
-			return x, i
-
-
-def assembly_stats(genome):
-	'''Returns a dict containing genome stats'''
-	seq_records = SeqIO.parse(genome, "fasta")
-	lengths = sorted([len(x.seq) for x in seq_records], reverse=True)
-	stats = {
-		'Sequence count': len(lengths),
-		'Total length': sum(lengths),
-		'Max length': lengths[0]
-	}
-	stats['N50'], stats['L50'] = get_N_L_metric(lengths, 50)
-	stats['N90'], stats['L90'] = get_N_L_metric(lengths, 90)
-	return stats
-
-
-def write_assembly_stats(args, stats, prefix, s_format="human"):
-	'''Write assembly stats to a file'''
-	ext_dict = {"human": "txt", "json": "json", "table": "tsv"}
-	filename = f"{prefix}.assembly.{ext_dict[s_format]}"
-	with open(os.path.join(args.output_dir, filename), 'w') as dest:
-		if s_format == "human":
-			for k, v in stats.items():
-				print(f"{k}\t{v}", file=dest)
-		if s_format == "json":
-			print(json.dumps(stats), file=dest)
-		if s_format == "table":
-			header = "\t".join(stats.keys())
-			data = "\t".join(list(map(str, stats.values())))
-			print(f"#{header}", file=dest)
-			print(f"{data}", file=dest)
-	return filename
 
 
 def check_last_step(args, step):
@@ -1011,7 +616,8 @@ def main():
 		if args.force:
 			shutil.rmtree(args.output_dir)
 		else:
-			logger.critical("Output directory \"%s\" already exists. Use --force to overwrite.",
+			logger.critical(
+				"Output directory \"%s\" already exists. Use --force to overwrite.",
 				args.output_dir)
 			raise FileExistsError(args.output_dir)
 	try:
@@ -1070,12 +676,6 @@ def main():
 		else:
 			estimated_genome_size = args.genome_size_estimation
 		args.genome = assemble(args, reads, estimated_genome_size)
-		stats = assembly_stats(args.genome)
-		logger.info("Assembly length: %s", stats['Total length'])
-		logger.info("Contig count: %s", stats['Sequence count'])
-		logger.info("N50: %s", stats['N50'])
-		write_assembly_stats(args, stats, prefix=args.assembler, s_format="table")
-
 		check_last_step(args, 3)
 
 	# Short read polishing is only meaningful for flye assembly